Investigation of four algorithms to multiply two large integer numbers and recording their running times.
One such algorithm for multipling two numbers is a naive looping addition of one number n
times. Another implements standard highschool multiplication, using shifting arrays. The
third uses recursion to calculate the product. The fourth and final algorithm uses a modified
version of the recursive algorithm to simplify and remove one multiplication step to decrease
the algoritm running time for large numbers. The TestLargeInteger class creates two large
integers, and separate executor services for each algorithm to record the running times of 1000
cycles for each. The running time results are then writen to a csv file that can be opened
in Excel.